requirements.txt
black==22.3.0
flask==2.1.1
Flask_APScheduler==1.12.3
Flask_Cors==3.0.10
flask_mongoengine==1.0.0
flask_restx==0.5.1
isodate==0.6.1
mongoengine==0.23.1
python-dotenv==0.20.0
requests==2.26.0
rsa==4.8
